# [#347] Migrate from Knative Configuration to Knative Service