--- a/packages/operator/controllers/modelroute_controller_test.go
+++ b/packages/operator/controllers/modelroute_controller_test.go
@@ -36,7 +36,7 @@ import (
 const (
 	mrName          = "test-mr"
 	mrURL           = "/test/url"
-	timeout         = time.Second * 300
+	timeout         = time.Second * 5
 	istioIngressSvc = "istio-ingressgateway.istio-system.svc.cluster.local"
 )