… helms/odahu-flow-core/templates/edge-gateway.yaml
… helms/odahu-flow-core/templates/edge-ingress.yaml
… helms/odahu-flow-core/templates/operator/generated/odahuflow.odahu.org_modeldeployments.yaml
… helms/odahu-flow-core/templates/service_catalog/virtualservice.yaml
… helms/odahu-flow-core/values.yaml
… packages/operator/controllers/modeldeployment_controller_test.go
… packages/operator/docs/docs.go
… packages/operator/docs/swagger.json
… packages/operator/docs/swagger.yaml
… packages/operator/go.mod
… packages/operator/hack/tests/thirdparty_crds/services.serving.knative.dev.yaml
… packages/operator/pkg/apiserver/routes/v1/deployment/model_deployment_test.go
… packages/sdk/odahuflow/sdk/models/model_deployment_status.py