Closes #347.
This PR adopts recommended way to use Knative - with Knative Service abstraction. What it does is automatically creates Knative Configuration **and** a rule (`VirtualService`) for Istio Ingress that forwards an incoming request to the latest ready Revision basing on the `Host` header. For example for KService `my-ml-model` in namespace `odahu-flow-deployment` it creates a rule that forwards requests to "my-ml-model.odahu-flow-deployment.example.com" to the pod that runs the latest revision. This hostname is avaible in KService status.
To route requests basing on request path we create another `VirtualService` that listens for requests with corresponding path prefix, removes this path prefix, adds a corresponding `Host` header and forwards the request to **itself, to the same exact Istio Ingress**. On the second routing iteration the request is handled by Knative's rules.
![image](https://user-images.githubusercontent.com/25623173/110021614-062c2c00-7d3c-11eb-94bc-411122fd10c3.png)

**Challenge with `X-Request-ID` header:** in case of "double" routing described above, Istio Ingress generates `X-Request-ID` header both times. That breaks the inference request-response logging, because tapping (collecting and loggging requests on ingress) happens on the first routing iteration. It remembers the first request ID. When the request gets to model it has the second request ID. And the second one returns to user in response. To fix it, I added an `EnvoyFilter` that prevents overriding `X-Request-ID` header on gateways. 

Additional sub-tasks:
- [X] Remove "edge" gateway and use the one created by Knative
- [X] Remove config related to "edge" gateway

Performance test was done with wine model and pretty big batch. 100 measures were made on both old and new implementations. Basically there's no major difference.
|       |    before, sec |     after, sec |
|:------|----------:|----------:|
| count | 100       | 100       |
| mean  |   2.89102 |   2.87684 |
| std   |   1.54696 |   1.56815 |
| min   |   2.04786 |   2.01864 |
| 25%   |   2.14841 |   2.1045  |
| 50%   |   2.26101 |   2.24131 |
| 75%   |   2.77788 |   2.80515 |
| max   |   8.53763 |   8.60008 |